feat. : image 빈 설정
